added basic connection with backend
the frontend and backend can now communicate hurray! Some things need to be changed so it doesn't only work on a local machine but that's a start.
Note: use `npm start` to run the app instead of `ng serve` so that it uses the proxy settings necessary to communicate with the backen